requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
boto3
openai
psycopg2-binary
python-dotenv
requests
